+++ history.md
@@ -1,4 +1,9 @@
 # 修改历史
+## 0.12.0 (2020-10-10)
+* feat: 接入 `yyl-config-types`
+* feat: 新增 `yh.initBeforeScript()`, `yh.initAfterScript()`
+* feat: 重新整理 test 用例
+
 ## 0.11.4 (2020-09-22)
 * fix: yh.parseConfig localserver.root 路径调整
 ## 0.11.3 (2020-09-22)